Добавит возможность переиспользовать одиноковую часть карточек: на странице фаворитов, главной и странице предложения
