+++ project/src/components/favorite-card/favorite-card.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
 import offerProp from '../props-validation/offer.prop';
+import CardInformation from '../card-information/card-information';
 
 export default function FavoriteCard(props) {
   const {offer} = props;
-  const {id, previewImage, price, isFavorite, isPremium, rating, title, type} = offer;
+  const {previewImage, isPremium} = offer;
 
   return (
     <article className="favorites__card place-card">
@@ -18,28 +18,7 @@ export default function FavoriteCard(props) {
         </a>
       </div>
       <div className="favorites__card-info place-card__info">
-        <div className="place-card__price-wrapper">
-          <div className="place-card__price">
-            <b className="place-card__price-value">&euro;{price}</b>
-            <span className="place-card__price-text">&#47;&nbsp;night</span>
-          </div>
-          <button className={`place-card__bookmark-button ${isFavorite ? 'place-card__bookmark-button--active': ''} button`} type="button">
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark"></use>
-            </svg>
-            <span className="visually-hidden">In bookmarks</span>
-          </button>
-        </div>
-        <div className="place-card__rating rating">
-          <div className="place-card__stars rating__stars">
-            <span style={{width: `${Math.round(100 / 5 * rating)}%`}}></span>
-            <span className="visually-hidden">Rating</span>
-          </div>
-        </div>
-        <h2 className="place-card__name">
-          <Link to={`/offer/${id}`}>{title}</Link>
-        </h2>
-        <p className="place-card__type">{type}</p>
+        <CardInformation offer={offer}/>
       </div>
     </article>
   );
