+++ project/src/components/near-place-card/near-place-card.jsx
@@ -1,44 +1,24 @@
 import React from 'react';
-import {Link, useHistory} from 'react-router-dom';
-import {AppRoute} from '../../const';
 import offerProp from '../props-validation/offer.prop';
+import CardInformation from '../card-information/card-information';
 
 export default function NearPlaceCard(props) {
   const {offer} = props;
-  const {id, previewImage, price, isFavorite, rating, title, type} = offer;
-
-  const history = useHistory();
+  const {previewImage, isPremium} = offer;
 
   return (
     <article className="near-places__card place-card">
+      {isPremium &&
+        <div className="place-card__mark">
+          <span>Premium</span>
+        </div>}
       <div className="near-places__image-wrapper place-card__image-wrapper">
         <a href="#">
           <img className="place-card__image" src={previewImage} width="260" height="200" alt="Place image"/>
         </a>
       </div>
       <div className="place-card__info">
-        <div className="place-card__price-wrapper">
-          <div className="place-card__price">
-            <b className="place-card__price-value">&euro;{price}</b>
-            <span className="place-card__price-text">&#47;&nbsp;night</span>
-          </div>
-          <button className={`place-card__bookmark-button ${isFavorite ? 'place-card__bookmark-button--active': ''} button`} onClick={() => history.push(AppRoute.SIGN_IN)} type="button">
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark"></use>
-            </svg>
-            <span className="visually-hidden">To bookmarks</span>
-          </button>
-        </div>
-        <div className="place-card__rating rating">
-          <div className="place-card__stars rating__stars">
-            <span style={{width: `${Math.round(100 / 5 * rating)}%`}}></span>
-            <span className="visually-hidden">Rating</span>
-          </div>
-        </div>
-        <h2 className="place-card__name">
-          <Link to={`/offer/${id}`}>{title}</Link>
-        </h2>
-        <p className="place-card__type">{type}</p>
+        <CardInformation offer={offer}/>
       </div>
     </article>
   );
